+++ asphalt/sqlalchemy/component.py
@@ -119,7 +119,7 @@ class SQLAlchemyComponent(Component):
                     await call_in_executor(session.commit, executor=self.commit_executor)
             finally:
                 del session.info['ctx']
-                session.close()
+                await call_in_executor(session.close, executor=self.commit_executor)
 
         session = factory(info={'ctx': ctx})
         ctx.add_teardown_callback(teardown_session, pass_exception=True)